API: Reorder schema macro structure disignated initialisers.
C++ finally has designated initialisers but they still require
member order to match the definition, even though the point of
designated initilisers is to allow you to not care about order
and be explicit about specific members' values.

Reorder the members in the schema building macros to allow
them to be used from a modern enough C++ version.