Add Meson build system support
Adds support for Meson build system as an alternative to the Makefile.
This adds more flexibility for build setups that may not be the same as the
makefile expects and also allows for example simpler integration into Buildroot.
Changes to the source files were kept to the minimum, only parametrizing
some hardcoded paths.
If this commit is accepted, the project could also be added to the Meson WrapDB
as a next step.

Signed-off-by: Daniel Brát <danek.brat@gmail.com>