API: Add support for default values for numeric fields
Default value upport is available for integer (both signed and un),
floating-point, enumeration and bitfields.

The library will keep working correctly with any code compiled with
a prior version, or new code that includes only the cyaml.h header.
Code that includes the cyaml_ex.h will include the option to set up
a default value for loading said types; if the values are not found
in the yaml file the default value will be set inside the allocated
buffer.